refactor: created vacancy-labels enums
