+++ src/entities/vacancy/index.ts
@@ -1,5 +1,10 @@
-export { getVacancyTagLabels } from './lib/getVacancyTagLabels';
+export {
+  vacancyEmploymentTypeLabels,
+  vacancyPositionLabels,
+  vacancyWorkTypeLabels,
+} from './config/vacancy-labels';
 export { useVacancies } from './model/hooks/useVacancies';
 export { useVacancy } from './model/hooks/useVacancy';
 export type { Vacancy } from './model/types/vacancy.types';
+export * from './model/types/vacancy-enum.types';
 export { VacancyCard } from './ui/vacancy-card/VacancyCard';
