--- src/entities/vacancy/lib/getVacancyTagLabels.ts
@@ -1,21 +1,0 @@
-import {
-  getVacancyEmploymentTypeLabel,
-  getVacancyPositionLabel,
-  getVacancyWorkTypeLabel,
-} from '../model/constants/vacancy-labels';
-
-type VacancyTagFields = {
-  position: string;
-  employmentType: string;
-  workType: string;
-};
-
-export const getVacancyTagLabels = ({
-  position,
-  employmentType,
-  workType,
-}: VacancyTagFields): string[] => [
-  getVacancyPositionLabel(position),
-  getVacancyEmploymentTypeLabel(employmentType),
-  getVacancyWorkTypeLabel(workType),
-];
